% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{align_to_ref}
\alias{align_to_ref}
\title{Align EffectAllele to always be the reference genome allele}
\usage{
align_to_ref(dset, ref = c("REF_38", "REF_37"))
}
\arguments{
\item{dset}{object createed by \code{\link[arrow:open_dataset]{arrow::open_dataset()}} or \code{\link[dplyr:reexports]{dplyr::tibble()}}}

\item{ref}{Reference genome allele to align to. Varies in ~0.5\% of locations}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} or arrow query depending on whether dset is a tibble or arrow dataset
}
\description{
\code{\link[=align_to_ref]{align_to_ref()}} will:
\enumerate{
\item Filter any variants where EffectAllele or OtherAllele is not the reference genome allele
\item Flip EffectAllele such that it is always the reference genome allele
\item If OtherAllele is the reference genome allele, direction of B,Z and EAF are flipped.
}
}
\examples{
\dontrun{
align_to_ref(tidygwas_df, ref = "REF_38")
}
}
