% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{by_chrom}
\alias{by_chrom}
\title{Meta analyse a chromosome, one at a time!}
\usage{
by_chrom(ds, chrom, ref)
}
\arguments{
\item{ds}{an \code{\link[arrow:open_dataset]{arrow::open_dataset()}} object}

\item{chrom}{chromosome}

\item{ref}{reference allele}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Meta analyse a chromosome, one at a time!
}
\examples{
\dontrun{
by_chrom(ds, chrom = 1, ref = "REF_38")
}
}
