% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CAD_BRAP}
\alias{CAD_BRAP}
\title{CAD GWAS summary data within +/-1MB from each side of BRAP gene after proper re-formatting and quality control}
\format{
An object of class data.frame.
}
\source{
\href{https://www.cardiogramplusc4d.org/data-downloads/}
}
\usage{
data(CAD_BRAP)
}
\description{
CAD GWAS summary data within +/-1MB from each side of BRAP gene after proper re-formatting and quality control
}
\keyword{CAD}
