% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-gwas-catalog.R
\name{check_rest_avail}
\alias{check_rest_avail}
\title{Check if API access is available for a GWAS catalog study}
\usage{
check_rest_avail(study_id)
}
\arguments{
\item{study_id}{a study accession ID, e.g. "GCST000001"}
}
\value{
TRUE if the REST API is available for the study, FALSE otherwise
}
\description{
Check if API access is available for a GWAS catalog study
}
\examples{
\dontrun{
#' check_rest_avail("GCST000001")
}
}
