% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{create_lake}
\alias{create_lake}
\title{Create a data lake in hivestyle format}
\usage{
create_lake(dir, lake)
}
\arguments{
\item{dir}{a directory containing tidyGWAS cleaned summary statistics.
Each directory should contain a tidyGWAS_hivestyle directory.}

\item{lake}{a directory to create the data lake in.}
}
\description{
Create a data lake in hivestyle format
}
\examples{
\dontrun{
# create_lake("/path/to/dir", "/path/to/lake")
}
}
