% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated-meta-analysis.R
\name{deprec_meta_analyze}
\alias{deprec_meta_analyze}
\title{Perform meta-analysis of GWAS summary statistics datasets cleaned by tidyGWAS
This function is depreciated. Use \code{\link[=meta_analyse]{meta_analyse()}}
\code{\link[=deprec_meta_analyze]{deprec_meta_analyze()}} will:
\itemize{
\item flip the effect allele to the reference allele (using either GRCh37 or GRCh38), control with \code{ref}
\item variant id is constructed using \code{by} columns
\item EAF (allele frequency) and INFO (imputation quality) are weighted by sample size, if present
\item CaseN, N, ControlN and EffectiveN are all summed and carried forward
}}
\usage{
deprec_meta_analyze(
  dset,
  by = c("CHR", "POS_37", "RSID", "EffectAllele", "OtherAllele"),
  ref = c("REF_37", "REF_38")
)
}
\arguments{
\item{dset}{an \code{\link[arrow:open_dataset]{arrow::open_dataset()}} object}

\item{by}{a character vector of column names to group by. Default is c("CHR", "POS", "RSID", "EffectAllele", "OtherAllele")}

\item{ref}{either "REF_37" or "REF_38", depending on which column you want to use to standardize reference allele}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Perform meta-analysis of GWAS summary statistics datasets cleaned by tidyGWAS
This function is depreciated. Use \code{\link[=meta_analyse]{meta_analyse()}}
\code{\link[=deprec_meta_analyze]{deprec_meta_analyze()}} will:
\itemize{
\item flip the effect allele to the reference allele (using either GRCh37 or GRCh38), control with \code{ref}
\item variant id is constructed using \code{by} columns
\item EAF (allele frequency) and INFO (imputation quality) are weighted by sample size, if present
\item CaseN, N, ControlN and EffectiveN are all summed and carried forward
}
}
\examples{
\dontrun{
dset <- arrow::open_dataset("path_to/sumstats/")
res <- meta_analyze(dset)
}

}
