% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{detect_indels}
\alias{detect_indels}
\title{Detect Insertions/Deletions ('indels')}
\usage{
detect_indels(tbl, indel_strategy, filepaths, ...)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} as created by \code{\link[=parse_tbl]{parse_tbl()}}}

\item{indel_strategy}{Should indels be kept or removed?}

\item{filepaths}{a list of filepaths, created by \code{\link[=setup_pipeline_paths]{setup_pipeline_paths()}}}

\item{...}{pass additional arguments to \code{\link[arrow:read_delim_arrow]{arrow::read_delim_arrow()}}, if tbl is a filepath.}
}
\value{
a tbl
}
\description{
Indels are detected by examining \code{EffectAllele} and \code{OtherAllele}
}
\examples{
\dontrun{
detect_indels(sumstat, TRUE, filepaths = setup_pipeline_paths("testing"))
}
}
