% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGWAS.R
\name{download_ref_files}
\alias{download_ref_files}
\title{Download references files used by tidyGWAS}
\usage{
download_ref_files(save_dir)
}
\arguments{
\item{save_dir}{directory to save reference files to}
}
\description{
Download references files used by tidyGWAS
}
\examples{
\dontrun{
download_ref_files("path/to_dir")
}
}
