% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eQTL_BRAP_RAW}
\alias{eQTL_BRAP_RAW}
\title{Cis-eQTLs from the +/-1MB from the center of BRAP gene without any formatting and quality control}
\format{
An object of class data.frame.
}
\source{
\href{https://eqtlgen.org/cis-eqtls.html}
}
\usage{
data(eQTL_BRAP_RAW)
}
\description{
Cis-eQTLs from the +/-1MB from the center of BRAP gene without any formatting and quality control
}
\keyword{eQTL}
