% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.gwas.R
\name{extract.data}
\alias{extract.data}
\title{Fast extract a subset of GWAS data}
\usage{
\method{extract}{data}(
  data.path,
  data.type = c(".txt", ".txt.gz"),
  extract.by = c("ID", "Coord"),
  id.index,
  ref.id,
  chr.index,
  pos.index,
  chr,
  pos.lower,
  pos.upper
)
}
\arguments{
\item{data.path}{Path to the GWAS summary to extract}

\item{data.type}{To clarify whether the GWAS is a .txt file or compressed .txt file}

\item{extract.by}{Whether extract by RSID or the Coordinate of variants}

\item{id.index}{Column number of RSID in the GWAS summary data}

\item{ref.id}{A vector of RSID to match in the GWAS summary data}
}
\value{
A data frame that contains a subset of GWAS summary data
}
\description{
Fast extract a subset of the GWAS summary data set in reference to the ID provided.
The data set should be a .txt file separated by tab
}
