% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_utils.R
\name{flag_duplicates}
\alias{flag_duplicates}
\title{Find all rows which are part of a set of duplicated rows}
\usage{
flag_duplicates(
  tbl,
  column = c("rsid", "chr_pos", "chr_pos_ref_alt", "rsid_ref_alt")
)
}
\arguments{
\item{tbl}{a tibble with \code{\link[=tidyGWAS_columns]{tidyGWAS_columns()}}}

\item{column}{Which columns should be used to form a unique ID?}
}
\value{
a tibble with new columns dup_{column}
}
\description{
Many duplication tools such as \code{\link[base:duplicated]{base::duplicated()}} or \code{\link[dplyr:distinct]{dplyr::distinct()}}
identify rows which are duplications. It is often useful to see ALL rows
which are part of the duplication set, and not just the second row.

creates new column: \code{dup_rsid} or \code{dup_chr_pos}, a T/F flag.
Specifically, flags both rows in a duplication pair, and not just first or
last duplicate row, making it easy to work with all rows that are part of a
duplication
}
\examples{
\dontrun{

# will tag multi-allelics as duplications
flag_duplicates(tbl, column = "rsid")
flag_duplicates(tbl, column = "chr_pos")
# if you are interested in rows that are variant duplications
flag_duplicates(tbl, column = "rsid_ref_alt")
flag_duplicates(tbl, column = "chr_pos_ref_alt")

}
}
