% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_utils.R
\name{flag_invalid_rsid}
\alias{flag_invalid_rsid}
\title{Detect entries that are not valid rsID's in GWAS summary statistics}
\usage{
flag_invalid_rsid(tbl, regex = "^[rR][sS]?\\\\d{1,10}$")
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} with column \code{RSID}.}

\item{regex}{regex used to detect non-RSIDs}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with column \code{invalid_rsid}
}
\description{
Detect entries that are not valid rsID's in GWAS summary statistics
}
\examples{
\dontrun{
flag_invalid_rsid(tbl) |>
dplyr::filter(invalid_rsid)
}
}
