% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.qc.data.R
\name{format.data}
\alias{format.data}
\title{Convert a GWAS Summary data into a standard format for downstream colocalization analysis}
\usage{
\method{format}{data}(
  data = NULL,
  data.path = NULL,
  TRAIT,
  RSID = "RSID",
  CHR = "CHR",
  POS = "POS",
  EA = "EA",
  NEA = "NEA",
  EAF = NULL,
  MAF = NULL,
  BETA = "BETA",
  SE = "SE",
  P = NULL,
  LOG10P = NULL,
  N = "N"
)
}
\description{
Convert a GWAS Summary data into a standard format for downstream colocalization analysis
}
