% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-gwas-catalog.R
\name{from_gwas_catalog_region}
\alias{from_gwas_catalog_region}
\title{Query a specific region of interest for a using a gwas catalog study_id}
\usage{
from_gwas_catalog_region(study_id, chr, start, end)
}
\arguments{
\item{study_id}{a study accession ID, e.g. "GCST000001"}

\item{chr}{chromosome number, e.g. "1" - not "chr1"}

\item{start}{base pair start position, e.g. 1000000}

\item{end}{base pair end position, e.g. 2000000}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Query a specific region of interest for a using a gwas catalog study_id
}
\examples{
\dontrun{
#' get_gwas_catalog_region("GCST000001", "1", 1000000, 2000000)
}
}
