% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsgenome_wrappers.R
\name{get_bsgenome}
\alias{get_bsgenome}
\title{load SNPlocs and REF genome for GRCh 37 and 38}
\usage{
get_bsgenome()
}
\value{
a list of SNPlocs and BSgenome objects
}
\description{
load SNPlocs and REF genome for GRCh 37 and 38
}
\examples{
\dontrun{
bsgenome <- get_bsgenome()
}
}
