% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-targets.R
\name{get_open_targets_cs}
\alias{get_open_targets_cs}
\title{Query Open targets for all credible sets containing the variant}
\usage{
get_open_targets_cs(
  variant_id,
  page_size = 500,
  api_url = "https://api.platform.opentargets.org/api/v4/graphql"
)
}
\arguments{
\item{variant_id}{in the open targets format chr_pos_ref_alt: 5_100_101_A_C}

\item{page_size}{number of results per page, default 500}

\item{api_url}{URL of the Open Targets GraphQL API, default}
}
\value{
a \link[dplyr:reexports]{dplyr::tibble} with the following columns:
\itemize{
\item \code{CHR}: chromosome
\item \code{POS}: position
\item \code{P}: p-value
\item \code{REF}: reference allele
\item \code{ALT}: alternate allele
\item \code{description}: trait description
\item \code{study_id}: study ID
}
}
\description{
Query Open targets for all credible sets containing the variant
}
\examples{
\dontrun{
get_open_targets_cs("7_140459051_C_G")
}
}
