% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsgenome_wrappers.R
\name{get_ref_data}
\alias{get_ref_data}
\title{Read in RsMergeArch file}
\usage{
get_ref_data()
}
\value{
a tibble
}
\description{
Read in RsMergeArch file
}
\examples{
\dontrun{
rs_merge_arch <- get_ref_data()
}
}
