% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.data.R
\name{harmonise.coloc.data}
\alias{harmonise.coloc.data}
\title{To harmonise 2 GWAS summary data}
\usage{
harmonise.coloc.data(data.a, data.b)
}
\arguments{
\item{data.a}{The first gwas summary data set you want to harmonise. Trait, rsid, ea, and nea must be available in the gwas summary while eaf is optional.}

\item{data.b}{The second gwas summary data set you want to harmonise. Trait, rsid, ea, and nea must be available in the gwas summary while eaf is optional.}
}
\value{
A harmonised data with 2 gwas summary data.
}
\description{
Harmonise 2 GWAS summary data primarily for conventional coloc analysis.
By default, the RSID, EA, and NEA will be referred to the first data set.
}
