% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp_qc.R
\name{infer_build}
\alias{infer_build}
\title{Infer what genome build a GWAS summary statistics file is on.}
\usage{
infer_build(sumstat, n_snps = 10000)
}
\arguments{
\item{sumstat}{a data.frame with the columns "CHR" and "POS"}

\item{n_snps}{number of snps to check CHR and POS for}
}
\value{
either `"37"` or `"38"`
}
\description{
Infer what genome build a GWAS summary statistics file is on.
}
\examples{
\dontrun{
genome_build <- infer_build(gwas_sumstats)
}
}
