% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-ldsR.R
\name{ldsR-reexports}
\alias{ldsR-reexports}
\alias{celltype_analysis}
\alias{from_tidyGWAS}
\alias{get_annot_names}
\alias{ldsc_h2}
\alias{ldsc_rg}
\alias{munge}
\alias{partition_h2}
\title{Re-exports from ldsR}
\usage{
celltype_analysis(sumstat, covariate_dir, ldscore_dir, weights = NULL)

from_tidyGWAS(tbl, n = c("N", "EffectiveN"))

get_annot_names(ldscore_dir)

ldsc_h2(
  sumstat,
  pop_prev = NULL,
  sample_prev = 0.5,
  weights = NULL,
  M = NULL,
  n_blocks = 200
)

ldsc_rg(sumstats1, sumstats2, weights = NULL, M = NULL, n_blocks = 200)

munge(dset, info_filter = 0.9, eaf_filter = 0.01)

partition_h2(
  sumstat,
  ldscore_dirs,
  subset_annots = NULL,
  overlapping_annotations = FALSE,
  weights = NULL,
  n_blocks = 200
)
}
\description{
Re-exports from ldsR
}
\keyword{internal}
