% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_utils.R
\name{liability_h2}
\alias{liability_h2}
\title{Convert an estimate of observed-scale heritability to liability scale heritability}
\usage{
liability_h2(obs_h2, pop_prev, sample_prev = 0.5)
}
\arguments{
\item{obs_h2}{observed-scale heritability}

\item{pop_prev}{prevalence of the disorder in the general population}

\item{sample_prev}{the prevalence of the disorder in the sample.
Default value is 0.5, reflecting a case-control study using effective N as sample size}
}
\value{
a double
}
\description{
Convert an estimate of observed-scale heritability to liability scale heritability
}
\examples{
liability_h2(0.25, 0.02)
}
