% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{meta_analyse}
\alias{meta_analyse}
\title{Improved meta-analysis using tidyGWAS:ed files}
\usage{
meta_analyse(
  ds,
  chromosomes = c(1:22),
  min_EAF = NULL,
  ref = c("REF_38", "REF_37"),
  safe_mode = FALSE
)
}
\arguments{
\item{ds}{a dataset object, see \code{\link[arrow:open_dataset]{arrow::open_dataset()}}}

\item{chromosomes}{Which chrosomes to apply meta-analysis across. Default is autosomes: 1:22}

\item{min_EAF}{Filter on minimal EAF (effect allele frequency) value. (0 - 0.5)}

\item{ref}{Reference genome version to use for reference allele}

\item{safe_mode}{Apply additional filters to ensure no missing/inf/NaN across key columns?
This is set to FALSE by default because these filters should already be
guaranteed with the use of \code{\link[=tidyGWAS]{tidyGWAS()}}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
@description
\code{\link[=meta_analyse]{meta_analyse()}} will:
\itemize{
\item flip the effect allele to be the reference allele (using either GRCh37 or GRCh38), control with \code{ref}
\item variant id is constructed using RSID:EffectAllele:OtherAllele (Which is now RSID:REF:ALT)
\item EAF (allele frequency) and INFO (imputation quality) are weighted by sample size, if present
\item CaseN, N, ControlN and EffectiveN are all summed and carried forward
}
}
\examples{
\dontrun{
meta_analyse2(ds, min_EAF = 0.01)
}
}
