% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis_beta.R
\name{meta_analyse2}
\alias{meta_analyse2}
\title{Improved meta-analysis using tidyGWAS:ed files}
\usage{
meta_analyse2(
  ds,
  min_EAF = NULL,
  ref = c("REF_38", "REF_37"),
  chromosomes = c(1:22)
)
}
\arguments{
\item{ds}{a \code{\link[arrow:dataset]{arrow::dataset()}}}

\item{min_EAF}{Filter on minimal EAF (effect allele frequency) value. (0 - 0.5)}

\item{ref}{Reference genome version to use for reference allele}

\item{chromosomes}{Which chrosomes to apply meta-analysis across. Default is autosomes: 1:22}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Improved meta-analysis using tidyGWAS:ed files
}
\examples{
\dontrun{
meta_analyse2(ds, min_EAF = 0.01)
}
}
