% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{meta_analyze_by_chrom}
\alias{meta_analyze_by_chrom}
\title{meta_analyze summary statistics, one chromosome at a time!
This function is exposed to allow for testing using real data}
\usage{
meta_analyze_by_chrom(dset, chrom, by, ref)
}
\arguments{
\item{dset}{an \code{\link[arrow:open_dataset]{arrow::open_dataset()}} object}

\item{chrom}{chromosome to use for meta-analysis}

\item{by}{a character vector of column names to group by. Default is c("CHR", "POS", "RSID", "EffectAllele", "OtherAllele")}

\item{ref}{either "REF_37" or "REF_38", depending on which column you want to use to standardize reference allele}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
meta_analyze summary statistics, one chromosome at a time!
This function is exposed to allow for testing using real data
}
\examples{
\dontrun{
meta_analyze_by_crom(dset, chrom = "22")
}

}
