% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGWAS.R
\name{parse_tbl}
\alias{parse_tbl}
\title{Parse the input data.frame or filepath to tidyGWAS}
\usage{
parse_tbl(tbl, ...)
}
\arguments{
\item{tbl}{a \code{data.frame} or \code{character()} vector}

\item{...}{optional arguments passed to \code{\link[arrow:read_delim_arrow]{arrow::read_delim_arrow()}}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Parse the input data.frame or filepath to tidyGWAS
}
\examples{
\dontrun{
# read in file from disk, using tab delimiter and skipping first 100 rows
df <- parse_tbl("path/to_sumstats", delim = "\t", skip = 100)
df <- parse_tbl(tibble(RSID = "rs1001", B = 0.01, P = 0.005, SE = 0.001, N = 100))
}
}
