% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pQTL_BRAP_RAW}
\alias{pQTL_BRAP_RAW}
\title{Cis-pQTLs within +/- 1MB from each side of BRAP gene without any formatting and quality control}
\format{
An object of class data.frame.
}
\source{
\href{https://metabolomips.org/ukbbpgwas/}
}
\usage{
data(pQTL_BRAP_RAW)
}
\description{
Cis-pQTLs within +/- 1MB from each side of BRAP gene without any formatting and quality control
}
\keyword{pQTL}
