% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.qc.data.R
\name{qc.data}
\alias{qc.data}
\title{Quality Control the post-format standardized data}
\usage{
qc.data(data, option = "biallelic.snp", id = "rsid")
}
\description{
Quality Control the post-format standardized data
}
