% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicated rows}
\usage{
remove_duplicates(tbl, filepaths)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}}}

\item{filepaths}{a list of filepaths, created by \code{\link[=setup_pipeline_paths]{setup_pipeline_paths()}}}
}
\value{
a tbl
}
\description{
remove_duplicates uses either CHR:POS:EffectAllele:OtherAllele
or RSID:EffectAllele:OtherAllele to compute uniqueness.

If possible rows are arranged by p-value, to select the row with the smallest P.
}
\examples{
\dontrun{
paths <- setup_pipeline_paths("testing")
df <- remove_duplicates(sumstat, paths)
}
}
