% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{remove_rows_with_na}
\alias{remove_rows_with_na}
\title{Remove rows with missing values, and write out the removed files to disk}
\usage{
remove_rows_with_na(tbl, filepaths)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}}}

\item{filepaths}{a list of filepaths, created by \code{\link[=setup_pipeline_paths]{setup_pipeline_paths()}}}
}
\value{
a tbl
}
\description{
Remove rows with missing values, and write out the removed files to disk
}
\examples{
\dontrun{
df <- remove_rows_with_na(sumstat, setup_pipeline_paths("testing"))
}
}
