% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp_qc.R
\name{repair_chr_pos}
\alias{repair_chr_pos}
\title{Get CHR and POS using RSID}
\usage{
repair_chr_pos(sumstat)
}
\arguments{
\item{sumstat}{a dplyr::tibble with with atleast RSID, EffectAllele and OtherAllele}
}
\value{
a tibble
}
\description{
Get CHR and POS using RSID
}
\examples{
\dontrun{
sumstat_df <- repair_chr_pos(sumstat)
}

}
