% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp.R
\name{repair_ids}
\alias{repair_ids}
\title{Augment a data.frame with information from dbSNP}
\usage{
repair_ids(
  tbl,
  repair = c("rsid", "pos"),
  build = c("NA", "37", "38"),
  dbsnp_path
)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns CHR, POS, EffectAllele, OtherAllele
or RSID, EffectAllele, OtherAllele2.}

\item{repair}{"rsid" to repair RSID, "pos" to repair CHR and POS}

\item{build}{used if repair = "rsid" to specify genome build}

\item{dbsnp_path}{path to the directory containing the dbSNP dataset}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with added columns CHR, POS_37, POS_38 RSID, REF_37, ALT_37, REF_38, ALT_38, no_dbsnp_entry, incompat_alleles
}
\description{
Augment a data.frame with information from dbSNP
}
\examples{
\dontrun{
repair_ids(gwas_sumstats, repair = "rsid", build = "38", dbsnp_path = "/path/to/dbsnp")
}
}
