% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp_qc.R
\name{repair_rsid}
\alias{repair_rsid}
\title{Get RSID from either GRCh37 or GRCh38 reference genome, using CHR and POS}
\usage{
repair_rsid(sumstat, build)
}
\arguments{
\item{sumstat}{a tibble with CHR,POS, EffectAllele and OtherAllele}

\item{build}{Optional. Genome build, either 37 or 38. If not passed,
[infer_build()] will be called to get the genome build.}
}
\value{
a tibble
}
\description{
Get RSID from either GRCh37 or GRCh38 reference genome, using CHR and POS
}
\examples{
\dontrun{
sumstat_df <- repair_rsid(sumstat, bsgenome_list)
}

}
