% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair-stats.R
\name{repair_stats}
\alias{repair_stats}
\title{Repair statistics column in a GWAS summary statistics tibble}
\usage{
repair_stats(tbl, impute_freq = NULL, impute_n = FALSE)
}
\arguments{
\item{tbl}{a \code{data.frame} or \code{character()} vector}

\item{impute_freq}{Should allele frequency be imputed if it's missing?
Provide a filepath to a .parquet file with columns RSID, EffectAllele, OtherAllele, EAF.
where EAF corresponds to frequency of the EffectAllele.}

\item{impute_n}{Should N be imputed if it's missing? see discussion in:
https://github.com/GenomicSEM/GenomicSEM/wiki/2.1-Calculating-Sum-of-Effective-Sample-Size-and-Preparing-GWAS-Summary-Statistics}
}
\value{
a tibble
}
\description{
\code{repair_stats()} is a collection of functions that can be used to
infer missing columns in GWAS summary statistics. The functions are based on
functionality found online.
}
\examples{
\dontrun{
updated <- repair_stats(my_gwas)
}
}
