% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SBP_BRAP}
\alias{SBP_BRAP}
\title{SBP GWAS summary data within +/-1MB from each side of BRAP gene after proper re-formatting and quality control}
\format{
An object of class data.frame.
}
\source{
\href{https://www.nealelab.is/uk-biobank}
}
\usage{
data(SBP_BRAP)
}
\description{
SBP GWAS summary data within +/-1MB from each side of BRAP gene after proper re-formatting and quality control
}
\keyword{GWAS}
\keyword{data}
\keyword{summary}
