% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SBP_BRAP_RAW}
\alias{SBP_BRAP_RAW}
\title{SBP GWAS summary data within +/-1MB from each side of BRAP gene without any formatting and quality control}
\format{
An object of class data.frame.
}
\source{
\href{https://www.nealelab.is/uk-biobank}
}
\usage{
data(SBP_BRAP_RAW)
}
\description{
SBP GWAS summary data within +/-1MB from each side of BRAP gene without any formatting and quality control
}
\keyword{SBP}
