% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGWAS_helpers.R
\name{set_default_dbsnp_path}
\alias{set_default_dbsnp_path}
\title{Save the filepath for the dbSNP reference data}
\usage{
set_default_dbsnp_path(dbsnp_path)
}
\arguments{
\item{dbsnp_path}{filepath to dbSNP}
}
\description{
Save the filepath for the dbSNP reference data
}
\examples{
\dontrun{
set_default_dbsnp_path("data/local/dbSNP155")
}
}
