% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_names.R
\docType{data}
\name{ss_head_dict}
\alias{ss_head_dict}
\title{GWAS summary statistic header dictionary}
\format{
An object of class \code{list} of length 17.
}
\usage{
ss_head_dict
}
\description{
A list showing interpretations for a range of possible GWAS summary statistic column names.
}
\keyword{datasets}
