% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_utils.R
\name{strand_flip}
\alias{strand_flip}
\title{Strand flip alleles}
\usage{
strand_flip(tbl)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{EffectAllele} and \code{OtherAllele}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{EffectAllele} and \code{OtherAllele} flipped
}
\description{
Strand flip alleles
}
\examples{
\dontrun{
tbl <- strand_flip(tbl)
}
}
