% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_file}
\alias{test_file}
\title{PGC3 schizophrenia GWAS}
\format{
## `test_file`
A data frame with 100,000 rows and 12 columns:
\describe{
  \item{CHR}{Chromosome}
  \item{POS}{Genomic position}
  \item{RSID}{rsID from dbSNP}
  \item{EffecttAllele}{The allele that correspods to the effect, B}
  \item{OtherAllelle}{Non-effect allele}
  \item{B}{Effect, Beta}
  \item{SE}{Standard error of B}
  \item{EAF}{EffectAllele Frequency, frequency of EffectAllele}
  \item{INFO}{Imputation accuracy}
  \item{P}{Pvalue}
  \item{CaseN}{Number of cases}
  \item{ControlN}{Number of controls}
  ...
}
}
\source{
<https://figshare.com/ndownloader/files/34517828>
}
\usage{
test_file
}
\description{
A subset of data from the Psychiatics Genonomics Consortium GWAS on schizophrenia (european subset)
}
\keyword{datasets}
