% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGWAS-package.R
\docType{package}
\name{tidyGWAS-package}
\alias{tidyGWAS-package}
\title{tidyGWAS: Quality control for GWAS summary statistics with minimal filtering}
\description{
tidyGWAS employs a version of dbSNP transformed into Apache Arrow parquet files to restore and validate genome wide association summary statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ararder.github.io/tidyGWAS/}
}

}
\keyword{internal}
