% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGWAS.R
\name{tidyGWAS}
\alias{tidyGWAS}
\title{Clean GWAS summary statistics}
\usage{
tidyGWAS(tbl, bsgenome_objects, logfile = FALSE, name, outdir, ...)
}
\arguments{
\item{tbl}{a tibble or filepath}

\item{bsgenome_objects}{use get_bsgenome()}

\item{logfile}{Write messages to a logfile?}

\item{name}{name of the output directory}

\item{outdir}{where to copy results after finished}

\item{...}{arguments to other functions, such as study_n, build}
}
\value{
a tibble or NULL, depending on outdir
}
\description{
Clean GWAS summary statistics
}
\examples{
\dontrun{
tidyGWAS(tbl = "my_dataframe", logfile = "true", name = "test_run", outdir = "gwas_sumstat_dir")
}
}
