% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGWAS.R
\name{tidyGWAS_columns}
\alias{tidyGWAS_columns}
\title{Create a \code{\link[dplyr:reexports]{dplyr::tibble()}} with tidyGWAS column names}
\usage{
tidyGWAS_columns(
  tbl,
  CHR = "CHR",
  POS = "POS",
  RSID = "RSID",
  EffectAllele = "EffectAllele",
  OtherAllele = "OtherAllele",
  B = "B",
  SE = "SE",
  P = "P",
  EAF = "EAF",
  N = "N",
  CaseN = "CaseN",
  ControlN = "ControlN",
  INFO = "INFO",
  Z = "Z",
  OR = "OR"
)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} or something coercible to one}

\item{CHR}{chromosome}

\item{POS}{position}

\item{RSID}{rsID from dbSNP}

\item{EffectAllele}{allele corresponding to effect}

\item{OtherAllele}{non-effect allele}

\item{B}{Beta, effect,}

\item{SE}{standard error}

\item{P}{p value}

\item{EAF}{effect-allele frequency}

\item{N}{total sample size (case + control)}

\item{CaseN}{number of cases (for case-control phenotypes)}

\item{ControlN}{number of controls (for case-control phenotypes)}

\item{INFO}{INFO score, imputation accuracy}

\item{Z}{Z score}

\item{OR}{odds-ratio}
}
\value{
a tibble with changes column names
}
\description{
\code{\link[=tidyGWAS]{tidyGWAS()}} requires the column names to be in a specific format.
This function facilitates that format, and also provides a list of all the columns
tidyGWAS considers valid. Any columns with a non tidyGWAS column name will be dropped.
}
\details{
The function is a simple wrapper around \code{\link[dplyr:select]{dplyr::select()}}, with each possible
tidyGWAS column as an argument-
}
\examples{
wrong_format <- dplyr::tibble(CHROM = 1, bp = 1000, A1 = "C", A2 = "A", Effect = 0.05)
formatted <- tidyGWAS_columns(
wrong_format, CHR = "CHROM", POS = "bp",
EffectAllele = "A1", OtherAllele = "A2", B = "Effect"
)
# columns that are wrongly named and NOT passed to tidyGWAS_columns() are dropped
wrong_format <- dplyr::tibble(CHROM = 1, bp = 1000, A1 = "C", A2 = "A", Effect = 0.05)
tidyGWAS_columns(
wrong_format, CHR = "CHROM", POS = "bp",
EffectAllele = "A1"
)


}
