% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGWAS_helpers.R
\name{update_rsid}
\alias{update_rsid}
\title{Title}
\usage{
update_rsid(tbl, filepath, dbsnp_path)
}
\arguments{
\item{tbl}{a \code{data.frame} or \code{character()} vector}

\item{filepath}{file to write removed rows to}

\item{dbsnp_path}{filepath to the dbSNP155 directory (untarred dbSNP155.tar)}
}
\value{
a data.frame with updated RSID column
}
\description{
Title
}
\examples{
\dontrun{
update_rsid(tbl, "path/to/file.parquet", "path/to/dbsnp")
}
}
