% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_validators.R
\name{validate_rsid}
\alias{validate_rsid}
\title{Validate format of the RSID column in a GWAS summary statistics file}
\usage{
validate_rsid(tbl, verbose = FALSE, outpath)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} as created by \code{\link[=parse_tbl]{parse_tbl()}}}

\item{verbose}{Explain filters in detail?}

\item{outpath}{Filepath: Where to write rows with invalid RSID?}
}
\value{
a tbl
}
\description{
Validate format of the RSID column in a GWAS summary statistics file
}
\examples{
\dontrun{
validate_rsid(sumstat, "~/invalid_rsid.parquet")
}
}
