% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{detect_indels}
\alias{detect_indels}
\title{Detect Insertions/Deletions ('indels')}
\usage{
detect_indels(tbl, keep_indels, filepaths)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} as created by \code{\link[=parse_tbl]{parse_tbl()}}}

\item{keep_indels}{Should indels be kept?}

\item{filepaths}{a list of filepaths, created by \code{\link[=setup_pipeline_paths]{setup_pipeline_paths()}}}
}
\value{
a tbl
}
\description{
Indels are detected by examining \code{EffectAllele} and \code{OtherAllele}
}
\examples{
\dontrun{
detect_indels(sumstat, TRUE, filepaths = setup_pipeline_paths("testing"))
}
}
