% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_utils.R
\name{flag_duplicates}
\alias{flag_duplicates}
\title{Fast detection of duplicates in RSID or CHR and POS columns}
\usage{
flag_duplicates(tbl, column = "rsid")
}
\arguments{
\item{tbl}{a tibble with tidyGWAS column names}

\item{column}{construct id with either 'rsid' or 'chr_pos' or 'chr_pos_ref_alt' or 'rsid_ref_alt'}
}
\value{
a tibble
}
\description{
creates new column: `dup_rsid` or `dup_chr_pos`, a T/F flag.
Specifically, flags both rows in a duplication pair, and not just first or
last duplicate row, making it easy to work with all rows that are part of a
duplication
}
\examples{
\dontrun{
flag_duplicates(tbl, column = "chr_pos")
flag_duplicates(tbl, column = "crsid")

}
}
