% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas-catalog.R
\name{from_gwas_catalog}
\alias{from_gwas_catalog}
\title{Download summary statistics from GWAS catalog}
\usage{
from_gwas_catalog(study_id, quiet = FALSE)
}
\arguments{
\item{study_id}{A single character string with study ID, e.g. "GCST90475332"}

\item{quiet}{TRUE/FALSE - controls progress bar for downloading}
}
\value{
a filepath to downloaded summary statistics
}
\description{
Download summary statistics from GWAS catalog
}
\examples{
\dontrun{
from_gwas_catalog("GCST90475332")
}
}
