% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.data.R
\name{harmonise.coloc.data}
\alias{harmonise.coloc.data}
\title{To harmonise 2 GWAS summary data}
\usage{
harmonise.coloc.data(data.a, data.b)
}
\arguments{
\item{data.a}{The 1st gwas summary data you want to harmonise.
We recommend to format the gwas data with 'format.data' function in 'tidygwas' pcakge.
Column trait, rsid, ea, and nea must be available in the gwas summary while eaf is optional.}

\item{data.b}{The 2nd gwas summary data you want to harmonise.
We recommend to format the gwas data with 'format.data' function in 'tidygwas' pcakge.
Column trait, rsid, ea, and nea must be available in the gwas summary while eaf is optional.}
}
\value{
A harmonised data with 2 gwas summary data.
}
\description{
Harmonise 2 GWAS summary data primarily for conventional coloc analysis.
By default, the RSID, EA, and NEA will be referred to the first data set.
}
