% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp_qc.R
\name{infer_build}
\alias{infer_build}
\title{Infer what genome build a GWAS summary statistics file is on.}
\usage{
infer_build(tbl, n_snps = 10000, dbsnp_path)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} with column names in \code{\link[=tidyGWAS_columns]{tidyGWAS_columns()}} format.}

\item{n_snps}{number of snps to check CHR and POS for}

\item{dbsnp_path}{filepath to dbSNP files in .parquet format.}
}
\value{
either "37" or "38"
}
\description{
Infer what genome build a GWAS summary statistics file is on.
}
\examples{
\dontrun{
genome_build <- infer_build(gwas_sumstats)
}
}
