% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis2.R
\name{meta_analyse2}
\alias{meta_analyse2}
\title{Improved meta-analysis using tidyGWAS:ed files
\code{\link[=meta_analyse2]{meta_analyse2()}} will:
\itemize{
\item flip the effect allele to be the reference allele (using either GRCh37 or GRCh38), control with \code{ref}
\item variant id is constructed using RSID:EffectAllele:OtherAllele (Which is now RSID:REF:ALT)
\item EAF (allele frequency) and INFO (imputation quality) are weighted by sample size, if present
\item CaseN, N, ControlN and EffectiveN are all summed and carried forward
}}
\usage{
meta_analyse2(
  ds,
  min_EAF = NULL,
  ref = c("REF_38", "REF_37"),
  chromosomes = c(1:22)
)
}
\arguments{
\item{ds}{a dataset object, see \code{\link[arrow:open_dataset]{arrow::open_dataset()}}}

\item{min_EAF}{Filter on minimal EAF (effect allele frequency) value. (0 - 0.5)}

\item{ref}{Reference genome version to use for reference allele}

\item{chromosomes}{Which chrosomes to apply meta-analysis across. Default is autosomes: 1:22}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Improved meta-analysis using tidyGWAS:ed files
\code{\link[=meta_analyse2]{meta_analyse2()}} will:
\itemize{
\item flip the effect allele to be the reference allele (using either GRCh37 or GRCh38), control with \code{ref}
\item variant id is constructed using RSID:EffectAllele:OtherAllele (Which is now RSID:REF:ALT)
\item EAF (allele frequency) and INFO (imputation quality) are weighted by sample size, if present
\item CaseN, N, ControlN and EffectiveN are all summed and carried forward
}
}
\examples{
\dontrun{
meta_analyse2(ds, min_EAF = 0.01)
}
}
