% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicated rows from a summary statistics file}
\usage{
remove_duplicates(tbl, columns = NULL, filepath)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}}, with columns in \code{tidyGWAS()}}

\item{columns}{a character vector of columns passed to \code{dplyr::distinct()}}

\item{filepath}{a filepath to write out the removed rows}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with duplicates removed
}
\description{
Remove duplicated rows from a summary statistics file
}
\examples{
\dontrun{
remove_duplicates(tbl, c("CHR", "POS",), "duplicated_rows.tsv")
}
}
