% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp_qc.R
\name{repair_chr_pos}
\alias{repair_chr_pos}
\title{Get CHR and POS using RSID}
\usage{
repair_chr_pos(tbl, dbsnp_path)
}
\arguments{
\item{tbl}{a [dplyr::tibble()] with column names in [tidyGWAS_columns()] format.}

\item{dbsnp_path}{filepath to dbSNP files in .parquet format.}
}
\value{
a [dplyr::tibble()] with columns `CHR`, `POS`, `POS_37` and `CHR_37` added
}
\description{
Get CHR and POS using RSID
}
\examples{
\dontrun{
sumstat_df <- repair_chr_pos(sumstat)
}

}
