% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsgenome_wrappers.R
\name{repair_chr_pos}
\alias{repair_chr_pos}
\title{Get CHR and POS from a reference genome with rsID}
\usage{
repair_chr_pos(sumstat, bsgenome_objects, .filter_callback)
}
\arguments{
\item{sumstat}{a dplyr::tibble with with atleast RSID, EffectAllele and OtherAllele}

\item{bsgenome_objects}{a list containing BSgenome genoms and snp_locs. see get_bsgenome_objects}

\item{.filter_callback}{a function that can be run at the end.}
}
\value{
a tibble
}
\description{
Get CHR and POS from a reference genome with rsID
}
\examples{
\dontrun{
sumstat_df <- repair_chr_pos(sumstat, bsgenome_list)
}

}
