% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp_qc.R
\name{repair_rsid}
\alias{repair_rsid}
\title{Use CHR and POS to get RSID from dbSNP 155}
\usage{
repair_rsid(tbl, build = c("NA", "37", "38"), dbsnp_path)
}
\arguments{
\item{tbl}{a [dplyr::tibble()] with column names in [tidyGWAS_columns()] format.}

\item{build}{Genome build of sumstats. if `'NA'` [infer_build()] will be used
to automatically detect build.}

\item{dbsnp_path}{filepath to dbSNP files in .parquet format.}
}
\value{
a [dplyr::tibble()] with columns `CHR`, `POS`, `POS_37` and `CHR_37` added
}
\description{
This function assumes tidyGWAS column names, see [tidyGWAS_columns()]
}
\examples{
\dontrun{
sumstat_df <- repair_rsid(sumstat, bsgenome_list)
}

}
