% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp_qc.R
\name{repair_rsid}
\alias{repair_rsid}
\title{Use CHR and POS to get RSID from dbSNP 155}
\usage{
repair_rsid(sumstat, build = c("NA", "37", "38"))
}
\arguments{
\item{sumstat}{a tibble with CHR,POS, EffectAllele and OtherAllele}

\item{build}{Optional. Genome build, either NA, 37 or 38. If NA,
[infer_build()] will be called to get the genome build.}
}
\value{
a tibble
}
\description{
This function assumes tidyGWAS column names, see [tidyGWAS_columns()]
}
\examples{
\dontrun{
sumstat_df <- repair_rsid(sumstat, bsgenome_list)
}

}
