% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp_qc.R
\name{repair_rsid}
\alias{repair_rsid}
\title{Use CHR and POS to get RSID from dbSNP 155}
\usage{
repair_rsid(tbl, build = c("NA", "37", "38"), dbsnp_path)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} with column names in \code{\link[=tidyGWAS_columns]{tidyGWAS_columns()}} format.}

\item{build}{Genome build of sumstats. if \code{'NA'} \code{\link[=infer_build]{infer_build()}} will be used
to automatically detect build.}

\item{dbsnp_path}{filepath to dbSNP files in .parquet format.}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{CHR}, \code{POS}, \code{POS_37} and \code{CHR_37} added
}
\description{
This function assumes tidyGWAS column names, see \code{\link[=tidyGWAS_columns]{tidyGWAS_columns()}}
}
\examples{
\dontrun{
sumstat_df <- repair_rsid(sumstat, bsgenome_list)
}

}
