% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp_qc.R
\name{repair_rsid}
\alias{repair_rsid}
\title{Use CHR and POS to get RSID from dbSNP 155}
\usage{
repair_rsid(tbl, build = c("NA", "37", "38"), dbsnp_path)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}}, formated with \code{\link[=tidyGWAS_columns]{tidyGWAS_columns()}}}

\item{build}{Can be used to skip \code{\link[=infer_build]{infer_build()}}}

\item{dbsnp_path}{filepath to the dbSNP155 directory (untarred dbSNP155.tar)}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{CHR}, \code{POS}, \code{POS_37} and \code{CHR_37} added
}
\description{
This function assumes tidyGWAS column names, see \code{\link[=tidyGWAS_columns]{tidyGWAS_columns()}}
}
\examples{
\dontrun{
sumstat_df <- repair_rsid(sumstat, bsgenome_list)
}

}
