% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsgenome_wrappers.R
\name{repair_rsid}
\alias{repair_rsid}
\title{Get RSID from either GRCh37 or GRCh38 reference genome, using CHR and POS}
\usage{
repair_rsid(sumstat, bsgenome_objects, build, .filter_callback)
}
\arguments{
\item{sumstat}{a dplyr::
tibble with atleast CHR,POS, EffectAllele and OtherAllele}

\item{bsgenome_objects}{a list containing BSgenome genoms and snp_locs. see get_bsgenome_objects}

\item{build}{genome build, either '37' or '38'}

\item{.filter_callback}{a function that will be run at the end, with the dataframe as input}
}
\value{
a tibble
}
\description{
Get RSID from either GRCh37 or GRCh38 reference genome, using CHR and POS
}
\examples{
\dontrun{
sumstat_df <- repair_rsid(sumstat, bsgenome_list)
}

}
