% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_utils.R
\name{repair_stats}
\alias{repair_stats}
\title{Repair statistics column in a GWAS summary statistics tibble}
\usage{
repair_stats(tbl)
}
\arguments{
\item{tbl}{a \code{data.frame} or \code{character()} vector}
}
\value{
a tibble
}
\description{
\code{repair_stats()} is a collection of functions that can be used to
infer missing columns in GWAS summary statistics. The functions are based on
functionality found online.
}
\examples{
\dontrun{
updated <- repair_stats(my_gwas)
}
}
