% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair-stats.R
\name{repair_stats}
\alias{repair_stats}
\title{Repair statistics column in a GWAS summary statistics tibble}
\usage{
repair_stats(
  tbl,
  dbsnp_path,
  impute_freq = c("None", "EUR", "AMR", "AFR", "SAS", "EAS"),
  impute_freq_file = NULL,
  impute_n = FALSE
)
}
\arguments{
\item{tbl}{a \code{data.frame} or \code{character()} vector}

\item{dbsnp_path}{filepath to the dbSNP155 directory}

\item{impute_freq}{one of c("None", "EUR", "AMR", "AFR", "SAS", "EAS"). If None, no imputation is done.
Otherwise precomputed alleles frequence from 1000KG, selected ancestry is used}

\item{impute_freq_file}{filepath to a .parquet file with custom allele frequencies.
The file needs to be a tabular dataframe with columns RSID, EffectAllele, OtherAllele, EAF.
EAF should correspond to the frequency of the EffectAllele.}

\item{impute_n}{Should N be imputed if it's missing?}
}
\value{
a tibble
}
\description{
\code{repair_stats()} is a collection of functions that can be used to
infer missing columns in GWAS summary statistics. The functions are based on
functionality found online.
}
\examples{
\dontrun{
updated <- repair_stats(my_gwas)
}
}
