% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_utils.R
\name{repair_stats}
\alias{repair_stats}
\title{Repair statistics column in a GWAS summary statistics tibble}
\usage{
repair_stats(tbl)
}
\arguments{
\item{tbl}{input a tibble (data frame) of a GWAS sumamry statistics.}
}
\value{
a tibble
}
\description{
Requires the tidyGWAS column format
}
\examples{
\dontrun{
updated <- repair_stats(my_gwas)
}
}
