% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_utils.R
\name{repair_stats}
\alias{repair_stats}
\title{Repair statistics column in a GWAS summary statistics tibble}
\usage{
repair_stats(tbl)
}
\arguments{
\item{tbl}{input a tibble (data frame)}
}
\value{
a tibble
}
\description{
`tidyGWAS::repar_stats()` is a collection of functions that can be used to
infer missing columns in GWAS summary statistics. The functions are based on
functionality found online.
}
\details{
`repair_stats()` uses the tidyGWAS column specification. The functionality
has been tested assuming the input data.frame is a tibble.
}
\examples{
\dontrun{
updated <- repair_stats(my_gwas)
}
}
