% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_utils.R
\name{repair_stats}
\alias{repair_stats}
\title{Repair statistics column in a GWAS summary statistics tibble}
\usage{
repair_stats(tbl, verbose = FALSE)
}
\arguments{
\item{tbl}{a [dplyr::tibble()] with column names in [tidyGWAS_columns()] format.}

\item{verbose}{Should repair_stats print a masthead explaining what it does?}
}
\value{
a tibble
}
\description{
`repair_stats()` is a collection of functions that can be used to
infer missing columns in GWAS summary statistics. The functions are based on
functionality found online.
}
\examples{
\dontrun{
updated <- repair_stats(my_gwas)
}
}
