% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGWAS_helpers.R
\name{select_correct_columns}
\alias{select_correct_columns}
\title{Remove all columns that do not follow tidyGWAS naming}
\usage{
select_correct_columns(tbl)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Remove all columns that do not follow tidyGWAS naming
}
\examples{
\dontrun{
sumstats <- select_correct_columns(sumstats)
}
}
