% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGWAS.R
\name{setup_pipeline_paths}
\alias{setup_pipeline_paths}
\title{Create the folder structure for tidyGWAS}
\usage{
setup_pipeline_paths(name)
}
\arguments{
\item{name}{name of the output directory}
}
\value{
a list of filepaths
}
\description{
Create the folder structure for tidyGWAS
}
\examples{
setup_pipeline_paths("tidyGWAS_first")
}
