% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGWAS.R
\name{tidyGWAS}
\alias{tidyGWAS}
\title{Clean GWAS summary statistics}
\usage{
tidyGWAS(
  tbl,
  use_dbsnp = TRUE,
  name = stringr::str_replace_all(date(), pattern = c(` ` = "_", `:` = "_")),
  repair_cols = TRUE,
  implementation = "arrow",
  outdir,
  logfile = FALSE,
  log_on_err = "tidyGWAS_logfile.txt",
  keep_indels = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{tbl}{a tibble or filepath}

\item{use_dbsnp}{use dbSNP to apply filters? Default is TRUE}

\item{name}{name of the output directory. Default is a concotonated call to Sys.time()}

\item{repair_cols}{Should any missing columns be repaired? Default is TRUE}

\item{implementation}{Use arrow or bsgenome as backend to interact with dbSNP?, default is arrow}

\item{outdir}{Where should results be saved after a succesful run? Default is tempdir()}

\item{logfile}{Direct messages to a logfile? Default is FALSE}

\item{log_on_err}{Optional. Can pass a filepath to copy the logfile to when the function exists.
This can be very useful if running not interactively, and want to make sure
the log file exists even if the function errors.}

\item{keep_indels}{Should indels be kept? Default is TRUE}

\item{verbose}{Explain filters in detail? Default is FALSE.}

\item{...}{arguments that will be passed to other functions: Currently supports
study_n, build, rs_merge_arch}
}
\value{
a tibble or NULL, depending on outdir
}
\description{
Clean GWAS summary statistics
}
\examples{
\dontrun{
tidyGWAS(tbl = "my_dataframe", logfile = "true", name = "test_run", outdir = "gwas_sumstat_dir")
}
}
