% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGWAS.R
\name{tidyGWAS}
\alias{tidyGWAS}
\title{Clean GWAS summary statistics}
\usage{
tidyGWAS(
  tbl,
  bsgenome_objects,
  logfile = FALSE,
  name,
  outdir,
  keep_indels = TRUE,
  verbose = FALSE,
  log_on_err = "tidyGWAS_log.txt",
  ...
)
}
\arguments{
\item{tbl}{a tibble or filepath}

\item{bsgenome_objects}{use get_bsgenome()}

\item{logfile}{Write messages to a logfile?}

\item{name}{name of the output directory}

\item{outdir}{where to copy results after finished}

\item{keep_indels}{Should indels be kept? Default is TRUE}

\item{verbose}{Should tidyGWAS tell you everything it's doing? default is FALSE}

\item{log_on_err}{a filepath, used if logfile = TRUE. If tidyGWAS throws an error, the logfile
is copied to that file. Useful is running non-interactively, in which case the logfile
will be deleted if tidyGWAS errors.}

\item{...}{arguments to other functions, such as study_n, build}
}
\value{
a tibble or NULL, depending on outdir
}
\description{
Clean GWAS summary statistics
}
\examples{
\dontrun{
tidyGWAS(tbl = "my_dataframe", logfile = "true", name = "test_run", outdir = "gwas_sumstat_dir")
}
}
