% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGWAS.R
\name{tidyGWAS}
\alias{tidyGWAS}
\title{Execute validation and quality control of GWAS summmary statistics}
\usage{
tidyGWAS(
  tbl,
  dbsnp_path,
  ...,
  column_names = NULL,
  output_format = c("hivestyle", "parquet", "csv"),
  output_dir = tempfile(),
  CaseN = NULL,
  ControlN = NULL,
  N = NULL,
  impute_freq = c("None", "EUR", "AMR", "AFR", "SAS", "EAS"),
  impute_freq_file = NULL,
  impute_n = FALSE,
  allow_duplications = FALSE,
  min_EAF = NULL,
  build = c("NA", "37", "38"),
  default_build = c("37", "38"),
  indel_strategy = c("keep", "remove"),
  convert_p = 2.225074e-308,
  repair_cols = TRUE,
  logfile = FALSE
)
}
\arguments{
\item{tbl}{a \code{data.frame} or \code{character()} vector}

\item{dbsnp_path}{filepath to the dbSNP155 directory}

\item{...}{pass additional arguments to \code{\link[arrow:read_delim_arrow]{arrow::read_delim_arrow()}}, if tbl is a filepath.}

\item{column_names}{a named list of column names:
\code{list(RSID = "SNP", POS = "BP")}}

\item{output_format}{How should the finished cleaned file be saved?
\itemize{
\item 'csv' corresponds to \code{\link[arrow:write_csv_arrow]{arrow::write_csv_arrow()}}
\item 'parquet' corresponds to \code{\link[arrow:write_parquet]{arrow::write_parquet()}}
\item 'hivestyle' corresponds to \code{\link[arrow:write_dataset]{arrow::write_dataset()}} split by \code{CHR}
}}

\item{output_dir}{filepath to a folder where tidyGWAS output will be stored.
The folder should not yet exist. Note that the default argument is \code{tempfile()},
meaning that tidyGWAS output will not be saved by default over R sessions.}

\item{CaseN}{manually input number of cases}

\item{ControlN}{manually input number of controls}

\item{N}{manually input sample size}

\item{impute_freq}{one of c("None", "EUR", "AMR", "AFR", "SAS", "EAS"). If None, no imputation is done.
Otherwise precomputed alleles frequence from 1000KG, selected ancestry is used}

\item{impute_freq_file}{filepath to a .parquet file with custom allele frequencies.
The file needs to be a tabular dataframe with columns RSID, EffectAllele, OtherAllele, EAF.
EAF should correspond to the frequency of the EffectAllele.}

\item{impute_n}{Should N be imputed if it's missing?}

\item{allow_duplications}{Should duplicated variants be allowed? Useful if the munged sumstats are QTL sumstats}

\item{min_EAF}{Apply a filter on allele frequency prior to applying the algorithm. Useful to speed up cleaning of very large files}

\item{build}{If you are sure of what genome build ('37' or '38'), can be used to skip \code{\link[=infer_build]{infer_build()}} and speed up computation}

\item{default_build}{If only RSID exists, the build cannot be inferred. Nonetheless,
tidyGWAS applies a filter on incompatible alleles with GRCh37/38. In such a case,
tidyGWAS needs to decide on which reference genome to compare alleles with.}

\item{indel_strategy}{Should indels be kept or removed?}

\item{convert_p}{What value should be used for when P-value has been rounded to 0?}

\item{repair_cols}{Should any missing statistical columns be repaired if possible? calls \code{\link[=repair_stats]{repair_stats()}} if TRUE}

\item{logfile}{Should messages be redirected to a logfile?}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
\code{tidyGWAS()} performs a set of validations on input colummns, repairs missing
columns, and can add missing CHR/POS or RSID. In addition, CHR and POS is
standardised to GRCh38, with coordinates on GRCh37 added in as well.

Briefly, \code{tidyGWAS()} updates RSID if possible using the
\href{https://ftp.ncbi.nih.gov/snp/latest_release/JSON/}{refsnp-merged} file
from dbSNP. Each inputed column is then validated and coerced to
the correct type.

If statistis such as \code{P}, \code{B} are missing, \code{tidyGWAS()} will attempt to impute
them if possible using \code{\link[=repair_stats]{repair_stats()}}

Standard column names are assumed, BEFORE inputting into the function. This is
a deliberate decision as automatic parsing of some important column names
can be ambiguous For example, in some sumstats, A1 referes to effect allele,
while other formats use A1 as non-effect allele.
}
\examples{
\dontrun{
tidyGWAS(tbl = "my_dataframe", logfile = "true", name = "test_run", outdir = "gwas_sumstat_dir")
}
}
