% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{update_rsid}
\alias{update_rsid}
\title{Update rsIDs from dbSNP that have been merged into other RSIDs}
\usage{
update_rsid(tbl, filepaths, dbsnp_path)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} as created by \code{\link[=parse_tbl]{parse_tbl()}}}

\item{filepaths}{a list of filepaths, created by \code{\link[=setup_pipeline_paths]{setup_pipeline_paths()}}}

\item{dbsnp_path}{filepath to dbSNP155 directory}
}
\value{
a tbl
}
\description{
Update rsIDs from dbSNP that have been merged into other RSIDs
}
\examples{
\dontrun{
update_rsid(sumstat, filepaths = setup_pipeline_paths("testing"), dbsnp_path = "~/dbSNP155")
}
}
