% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_validators.R
\name{validate_columns}
\alias{validate_columns}
\title{Check that values in GWAS summary statistics columns are correct}
\usage{
validate_columns(
  tbl,
  col = c("B", "SE", "EAF", "N", "Z", "P", "POS", "CHR", "EffectAllele", "OtherAllele"),
  verbose = TRUE,
  convert_p = 2.225074e-308
)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} as created by \code{\link[=parse_tbl]{parse_tbl()}}}

\item{col}{Which column to check values in?}

\item{verbose}{Explain filters in detail?}

\item{convert_p}{What value should be used for P = 0?}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}, with a column added named as invalid_{col}.
If you validate the "B" column, validate_columns will add a TRUE/FALSE column
named invalid_B to the input tibble.
}
\description{
\code{validate_columns()} does not remove any rows, but adds a TRUE/FALSE
flag for the specified column.
}
\examples{
\dontrun{
gwas_file <- validate_columns(
 tbl = gwas_file,
 col = "B",
 verbose = FALSE,
 # if you want to keep 0 pvalues as 0.
 convert_p = 0
)
dplyr::filter(gwas_file, invalid_P)
}

}
