% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_validators.R
\name{validate_rsid}
\alias{validate_rsid}
\title{Validate format of the RSID column in a GWAS summary statistics file}
\usage{
validate_rsid(tbl, filepath)
}
\arguments{
\item{tbl}{a \code{data.frame} or \code{character()} vector}

\item{filepath}{filepath to write out removed rows}
}
\value{
a tbl
}
\description{
Validate format of the RSID column in a GWAS summary statistics file
}
\examples{
\dontrun{
validate_rsid(sumstat, "~/invalid_rsid.parquet")
}
}
