% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_validators.R
\name{validate_sumstat}
\alias{validate_sumstat}
\title{Validate statistics columns in a GWAS summary statistics file}
\usage{
validate_sumstat(tbl, remove_cols = c(""), filter_func, convert_p)
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}}}

\item{remove_cols}{Columns that should not be validated}

\item{filter_func}{handles reporting and writing removed files to disk}

\item{convert_p}{What value should be used for when P-value has been rounded to 0?}
}
\value{
a tbl
}
\description{
Validate statistics columns in a GWAS summary statistics file
}
\examples{
\dontrun{
validate_sumstat(sumstat, remove_cols = "EffectAllele", convert_p = 0)
}

}
