% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGWAS.R
\name{validate_with_dbsnp}
\alias{validate_with_dbsnp}
\title{Update CHR/POS/RSID/EffectAllele/OtherAllele for GWAS sumstats using dbSNP}
\usage{
validate_with_dbsnp(
  data_list,
  build = c("NA", "37", "38"),
  filepaths,
  dbsnp_path
)
}
\arguments{
\item{data_list}{a list containing three \code{\link[dplyr:reexports]{dplyr::tibble()}}s: main, indel, without_rsid}

\item{build}{Genome build of sumstats. if \code{'NA'} \code{\link[=infer_build]{infer_build()}} will be used
to automatically detect build.}

\item{filepaths}{a \code{list()} of filepaths, as created by \code{\link[=setup_pipeline_paths]{setup_pipeline_paths()}}}

\item{dbsnp_path}{filepath to dbSNP files in .parquet format.}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Update CHR/POS/RSID/EffectAllele/OtherAllele for GWAS sumstats using dbSNP
}
\examples{
\dontrun{

validate_with_dbsnp(sumstats, build = "NA", dbsnp_path = "/dbsnp155/dbsnp")
}

}
