% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyGWAS.R
\name{validate_with_dbsnp}
\alias{validate_with_dbsnp}
\title{Update CHR/POS/RSID/EffectAllele/OtherAllele for GWAS sumstats using dbSNP}
\usage{
validate_with_dbsnp(tbl, build = c("NA", "37", "38"), dbsnp_path, filter_func)
}
\arguments{
\item{tbl}{a [dplyr::tibble()] with column names in [tidyGWAS_columns()] format.}

\item{build}{Genome build of sumstats. if `'NA'` [infer_build()] will be used
to automatically detect build.}

\item{dbsnp_path}{filepath to dbSNP files in .parquet format.}

\item{filter_func}{used to remove rows which fail validation. see [make_callback()]}
}
\value{
a [dplyr::tibble()]
}
\description{
Update CHR/POS/RSID/EffectAllele/OtherAllele for GWAS sumstats using dbSNP
}
\examples{
\dontrun{

validate_with_dbsnp(sumstats, build = "NA", dbsnp_path = "/dbsnp155/dbsnp")
}

}
