% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsgenome_wrappers.R
\name{verify_chr_pos_rsid}
\alias{verify_chr_pos_rsid}
\title{Check that CHR POS and RSID all make sense in a GWAS summary statistic}
\usage{
verify_chr_pos_rsid(sumstat, bsgenome_objects, build)
}
\arguments{
\item{sumstat}{sumstat in tibble format, tidyGWAS column names}

\item{bsgenome_objects}{use get_bsgenome()}

\item{build}{optional, can be used to skip the infer_build step}
}
\value{
a tibble
}
\description{
Check that CHR POS and RSID all make sense in a GWAS summary statistic
}
\examples{
\dontrun{
gwas <- tidyGWAS::test_file
bs <- get_bsgenome()
# make_callback can be passed a filepath to write out removed rows to.
callback <- make_callback("~/output_folder/verify_chr_pos_rsid_removed_rows.tsv")
verify_chr_pos_rsid(gwas, bs, build = 37)
}
}
