% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp_qc.R
\name{verify_chr_pos_rsid}
\alias{verify_chr_pos_rsid}
\title{Compare CHR, POS and RSID with dbSNP reference data}
\usage{
verify_chr_pos_rsid(sumstat, build)
}
\arguments{
\item{sumstat}{sumstat in tibble format, tidyGWAS column names}

\item{build}{Optional. Genome build, either 37 or 38. If not passed,
[infer_build()] will be called to get the genome build.}
}
\value{
a tibble
}
\description{
Compare CHR, POS and RSID with dbSNP reference data
}
\examples{
\dontrun{
gwas <- tidyGWAS::test_file
bs <- get_bsgenome()
# make_callback can be passed a filepath to write out removed rows to.
callback <- make_callback("~/output_folder/verify_chr_pos_rsid_removed_rows.tsv")
verify_chr_pos_rsid(gwas, bs, build = 37)
}
}
