% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp_qc.R
\name{verify_chr_pos_rsid}
\alias{verify_chr_pos_rsid}
\title{Compare CHR, POS and RSID with dbSNP reference data}
\usage{
verify_chr_pos_rsid(tbl, build = c("NA", "37", "38"), dbsnp_path)
}
\arguments{
\item{tbl}{a `data.frame` or `character()` vector}

\item{build}{Can be used to skip [infer_build()]}

\item{dbsnp_path}{filepath to the dbSNP155 directory (untarred dbSNP155.tar)}
}
\value{
a [dplyr::tibble()] with columns `CHR`, `POS`, `POS_37` and `CHR_37` added
}
\description{
Compare CHR, POS and RSID with dbSNP reference data
}
\examples{
\dontrun{
gwas <- tidyGWAS::test_file
# make_callback can be passed a filepath to write out removed rows to.
callback <- make_callback("~/output_folder/verify_chr_pos_rsid_removed_rows.tsv")
verify_chr_pos_rsid(gwas, bs, build = 37)
}
}
