% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsnp_qc.R
\name{verify_chr_pos_rsid}
\alias{verify_chr_pos_rsid}
\title{Compare CHR, POS and RSID with dbSNP reference data}
\usage{
verify_chr_pos_rsid(sumstat, build = c("NA", "37", "38"))
}
\arguments{
\item{sumstat}{sumstat in tibble format, [tidyGWAS_columns()] names}

\item{build}{genome build. if unknown ("NA"), build will be
inferred using [infer_build()]}
}
\value{
a tibble
}
\description{
Compare CHR, POS and RSID with dbSNP reference data
}
\examples{
\dontrun{
gwas <- tidyGWAS::test_file
# make_callback can be passed a filepath to write out removed rows to.
callback <- make_callback("~/output_folder/verify_chr_pos_rsid_removed_rows.tsv")
verify_chr_pos_rsid(gwas, bs, build = 37)
}
}
